requirements.txt
python-dotenv==1.0.1
openai==1.12.0
pyarrow==17.0.0
pandas==2.2.2
langchain==0.1.9
langchain-openai==0.0.8
langchain_experimental==0.0.52
tabulate==0.9.0
Werkzeug==2.2.2
flask==2.2.2
gunicorn==20.1.0
